#38 start

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken')
 const { randomBytes } = require('crypto')
 const { promisify } = require('util')
 const { transport, makeEmail } = require('../mail')
+const { hasPermissions } = require('../utils')
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -18,7 +19,7 @@ const Mutations = {
             id: ctx.request.userId,
           },
         },
-      }
+      },
     }, info)
 
     return item
@@ -40,9 +41,13 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`)
+    const item = await ctx.db.query.item({ where }, `{id title user}`)
     // check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+    if (!ownsItem && hasPermissions) {
+      throw new Error("You don't have permission to delete")
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
@@ -148,6 +153,28 @@ const Mutations = {
       maxAge: 100 * 60 * 60 * 24 * 365,
     })
     // 8. return the new user
+    return updatedUser
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!')
+    }
+    const user = await ctx.db.query.user({
+      where: {
+        id: ctx.request.userId,
+      },
+    }, info)
+    hasPermissions(user, ['ADMIN', 'PERMISSIONUPDATE'])
+    const updatedUser = await ctx.db.mutation.updateUser({
+      where: {
+        id: args.userId,
+      },
+      data: {
+        permissions: {
+          set: args.permissions,
+        },
+      },
+    }, info)
     return updatedUser
   },
 }
